requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.11.3'
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2024.2.2; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
crontab==1.0.1
django==4.2.13; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-rq==2.10.2
django-scheduler==0.10.1; python_version >= '3.7'
freezegun==1.5.1; python_version >= '3.7'
icalendar==5.0.12; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
redis==5.0.4; python_version >= '3.7'
rq==1.16.2; python_version >= '3.7'
rq-scheduler==0.13.1
sentry-sdk==2.3.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version < '3.11'
urllib3==2.2.1; python_version >= '3.8'
